many, no k>=0

+++ internal/fischer/manual/A046042.java
@@ -21,8 +21,8 @@ public class A046042 extends GeneralizedEulerTransform {
   }
 
   @Override
-  protected int advanceH(final int k) {
-    return mSeqH.next().intValue();
+  protected long advanceH(final long k) {
+    return mSeqH.next().longValue();
   }
 
 }
